RecipeUpdater/web/recipe-backend/requirements.txt
azure-common==1.1.27
azure-core==1.17.0
azure-digitaltwins-core==1.1.0
azure-identity==1.6.0
azure-iot-device==2.7.1
azure-iot-hub==2.5.0
azure-keyvault-secrets==4.3.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
deprecation==2.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
janus==0.4.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msal==1.13.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
packaging==21.0
paho-mqtt==1.5.1
portalocker==1.7.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
uamqp==1.4.1
urllib3==1.26.6
vext==0.7.6
vext.gi==0.7.4
Werkzeug==2.0.1
